0006903: Moved getDataMap(), getDataObjectMap(), and dependent methods from SymmetricUtils to DataUtils
